notebooks fixed
