--- sbi/inference/__init__.py
@@ -1,3 +1,0 @@
-from sbi.inference.apt import APT
-from sbi.inference.sre import SRE
-from sbi.inference.snl import SNL
