+++ sbi/inference/snl/__init__.py
@@ -1,0 +1,1 @@
+from sbi.inference.snl.snl import SNL
