+++ sbi/inference/snpe/__init__.py
@@ -1,0 +1,4 @@
+from sbi.inference.snpe.base_snpe import base_snpe
+from sbi.inference.snpe.snpe_a import SNPE_A
+from sbi.inference.snpe.snpe_b import SNPE_B
+from sbi.inference.snpe.snpe_c import APT
