+++ sbi/inference/sre/__init__.py
@@ -1,0 +1,1 @@
+from sbi.inference.sre.sre import SRE
